+++ procrunner/__init__.py
@@ -53,7 +53,7 @@ from threading import Thread
 
 __author__ = """Markus Gerstel"""
 __email__ = "scientificsoftware@diamond.ac.uk"
-__version__ = "0.9.1"
+__version__ = "1.0.0"
 
 logger = logging.getLogger("procrunner")
 logger.addHandler(logging.NullHandler())